Update Metakit to v2.4.9.8.

+++ 8.x/mk/tcl/mk4tcl.h
@@ -1,5 +1,5 @@
 // mk4tcl.h --
-// $Id: mk4tcl.h 6125 2010-11-01 21:30:54Z patthoyts $
+// $Id$
 // This is part of Metakit, the homepage is http://www.equi4.com/metakit.html
 
 #include "mk4.h"
@@ -184,7 +184,7 @@ const c4_Property &AsProperty(Tcl_Obj *objPtr, const c4_View &view_);
 //  Cursors in Tcl are implemented as a pointer to an MkPath plus an index.
 
 MkPath &AsPath(Tcl_Obj *obj_);
-int &AsIndex(Tcl_Obj *obj_);
+long &AsIndex(Tcl_Obj *obj_);
 int SetCursorFromAny(Tcl_Interp *interp, Tcl_Obj *objPtr);
 
 // 24nov02: added to support releasing mutex lock during loop eval's
@@ -277,7 +277,7 @@ class MkTcl: public Tcl {
     ~MkTcl();
 
     c4_View asView(Tcl_Obj *obj_);
-    int &changeIndex(Tcl_Obj *obj_);
+    long &changeIndex(Tcl_Obj *obj_);
     c4_RowRef asRowRef(Tcl_Obj *obj_, int type_ = kExistingRow);
     int GetCmd();
     int SetValues(const c4_RowRef &row_, int objc, Tcl_Obj *const * objv);
